Refactors Database Tests into one Common class with all generic Database Tests and specific Database Tests
